pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.5"
python-dotenv = "^0.14.0"
requests = "^2.31.0"
pytest = "^8.1.1"
pytest-dotenv = "^0.5.2"
selenium = "^4.19.0"
gunicorn = "^22.0.0"
pymongo = "^4.9.1"
mongomock = "^4.2.0.post1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest-watch = "^4.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
